Move Relay Matrix address handler to avoid ftdi imports
