+++ src/fixate/core/switching.py
@@ -538,61 +538,6 @@ class PinValueAddressHandler(AddressHandler):
         print(f"0b{value:0{bits}b}")
 
 
-def _pins_for_one_relay_matrix(relay_matrix_num: int) -> list[Pin]:
-    """
-    A helper to create pin names for relay matrix cards.
-
-    Returns 16 pin names. If relay_matrix_num is 1:
-    1K1, 1K2, 1K3, ..., 1K16
-    """
-    return [f"{relay_matrix_num}K{relay}" for relay in range(1, 17)]
-
-
-class FTDIAddressHandler(PinValueAddressHandler):
-    """
-    An address handler which uses the ftdi driver to control pins.
-
-    We create this concrete address handler because we use it most
-    often. FT232 is used to bit-bang to shift register that are control
-    the switching in a jig.
-    """
-    def __init__(
-        self,
-        ftdi_description: str,
-        relay_matrix_count: int,
-        extra_pins: Sequence[Pin] = tuple()
-    ) -> None:
-        relay_matrix_pin_list = tuple(
-            itertools.chain.from_iterable(
-                _pins_for_one_relay_matrix(rm_number)
-                for rm_number in range(1, relay_matrix_count + 1)
-            )
-        )
-        self.pin_list = relay_matrix_pin_list + tuple(extra_pins)
-        # call the base class super _after_ we create the pin list
-        super().__init__()
-
-        # how many bytes? enough for every pin to get a bit. We might
-        # end up with some left-over bits. The +7 in the expression
-        # ensure we round up.
-        bytes_required = (len(self.pin_list) + 7) // 8
-        self._ftdi = ftdi.open(ftdi_description=ftdi_description)
-        self._ftdi.configure_bit_bang(
-            ftdi.BIT_MODE.FT_BITMODE_ASYNC_BITBANG,
-            bytes_required=bytes_required,
-            data_mask=4,
-            clk_mask=2,
-            latch_mask=1
-        )
-        self._ftdi.baud_rate = 115200
-
-    def close(self) -> None:
-        self._ftdi.close()
-
-    def _update_output(self, value: int) -> None:
-        self._ftdi.serial_shift_bit_bang(value)
-
-
 class VirtualAddressMap:
     """
     The supervisor loops through the attached virtual multiplexers each time a mux update is triggered.
